Update CS103.cpp
Have made changes to the savetoorderfunction so that it now saves order numbers and creates a txt file for each user and another txt file for admin. This was to get around the fix for users being able to view other users orders. Have also added the ability for users to delete their orders and admin to delete any users orders.